requirements.txt
tensorflow==2.0.0
opencv-python==4.1.1.26
bcolz==1.2.1
sklearn2
PyYAML
tqdm
